Running Time fixed
